% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{beta_update_joint}
\alias{beta_update_joint}
\title{Title : function for updating beta (jointly)}
\usage{
beta_update_joint(
  X,
  y,
  spt,
  cr_bt,
  cr_Sig,
  cr_f0,
  cr_tht,
  cr_bpr2,
  cr_btht,
  rho,
  linkfun,
  linkinv,
  mu.eta,
  mb,
  Sb
)
}
\arguments{
\item{X}{design matrix}

\item{y}{response variable}

\item{spt}{support of y}

\item{cr_bt}{current beta}

\item{cr_Sig}{variance-covariance matrix for current beta}

\item{cr_f0}{current f0 = \{f0(s_j), j=1(1)l\}}

\item{cr_tht}{current theta}

\item{cr_bpr2}{current \{second derivative of b(theta_i), i=1(1)n\}}

\item{cr_btht}{current b(theta)}

\item{rho}{MCMC update step size, a (scalar) in (0, 1]}
}
\description{
Title : function for updating beta (jointly)
}
\keyword{internal}
