% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspgldrmFit.R
\name{bspgldrmFit}
\alias{bspgldrmFit}
\title{Main MCMC function}
\usage{
bspgldrmFit(
  formula,
  data,
  X,
  y,
  link,
  mb,
  sb,
  dir_pr_parm,
  mu0,
  spt,
  init,
  bspgldrmControl,
  thetaControl
)
}
\description{
This function is called by the main \code{bspgldrm} function.
}
\keyword{internal}
