% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdpglm.R
\name{cdpglm}
\alias{cdpglm}
\title{Fit a Copula DP-GLM Model}
\usage{
cdpglm(
  formula,
  data = NULL,
  link = NULL,
  group_index,
  cdpglmControl = cdpglm.control(),
  thetaControl = theta.control()
)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}}.}

\item{data}{Optional data frame containing the variables in \code{formula}.}

\item{link}{Optional character string or link object specifying the mean link
function. If \code{NULL}, a default link is chosen based on the response
support: \code{"logit"} for responses in \code{(0, 1)}, \code{"log"} for
nonnegative responses, and \code{"identity"} otherwise.}

\item{group_index}{Vector specifying the group index for each observation.
Must have the same length as the response.}

\item{cdpglmControl}{Optional control arguments.
Passed as an object of class "dpglmControl", which is constructed by the
\code{dpglm.control} function.}

\item{thetaControl}{Optional control arguments for the theta update procedure.
Passed as an object of class "thetaControl", which is constructed by the
\code{theta.control} function.
See \code{theta.control} documentation for details.}
}
\value{
List of class "cdpglm" containing information including
posterior samples, Metropolis-Hastings acceptance rates, and model specification.
See details for more information.
}
\description{
Fit a Copula DP-GLM Model
}
\details{
The "cdpglm" class is a list of the following items.
\itemize{
\item \code{samples} A list containing the MCMC samples for
\code{beta}, \code{crm}, and \code{rho}.
\item \code{mb} Prior mean for \code{beta}.
\item \code{Sb} Prior variance-covariance matrix for \code{beta}.
\item \code{formula} Model formula.
\item \code{data} Model data frame.
\item \code{link} Link function. If a character string was passed to the
\code{link} argument, then this will be an object of class "link-glm".
Otherwise, it will be the list of three functions passed to the \code{link} argument.
\item \code{spt} Support of response variable \code{y} used in model fitting.
\item \code{mu0} Mean of the reference distribution obtained by model fit.
\item \code{burnin} The number of burn-in MCMC iterations.
\item \code{thin} Thinning parameter used for MCMC
\item \code{save} The number of saved MCMC iterations.
\item \code{rho_prior_shape} Prior parameters for dependence parameter \code{rho}.
\item \code{beta_acceptance} Proportion of accepted proposals for beta during MCMC.
\item \code{crm_acceptance} Proportion of accepted proposals for CRM during MCMC.
\item \code{rho_acceptance} Proportion of accepted proposals for rho during MCMC.
}
}
\examples{
\dontrun{
set.seed(123)

n_group <- 10
n_per_group <- 5
n <- n_group * n_per_group

id <- rep(seq_len(n_group), each = n_per_group)
x <- rnorm(n)
group_effect <- rep(rnorm(n_group, sd = 0.3), each = n_per_group)
y <- 1 + 0.5 * x + group_effect + rnorm(n, sd = 0.2)

dat <- data.frame(y = y, x = x, id = id)
fit <- cdpglm(
  y ~ x,
  data = dat,
  group_index = dat$id,
  link = "identity",
  cdpglmControl = cdpglm.control(
    spt = range(y),
    save = 50,
    thin = 1,
    burnin = 10,
    seed = 123
  )
 )
}

}
