% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdpglmFit.R
\name{cdpglmFit}
\alias{cdpglmFit}
\title{Main MCMC Fitting Function for CDPGLM}
\usage{
cdpglmFit(
  formula,
  data,
  X,
  y,
  group_index,
  link,
  spt,
  mu0,
  init,
  cdpglmControl,
  thetaControl
)
}
\arguments{
\item{formula}{Model formula supplied to \code{\link{cdpglm}}.}

\item{data}{Data frame containing the variables appearing in \code{formula}.}

\item{X}{Numeric design matrix.}

\item{y}{Numeric response vector.}

\item{group_index}{Vector identifying the dependence group for each observation.}

\item{link}{Link object containing \code{linkfun}, \code{linkinv}, and \code{mu.eta}.}

\item{spt}{Numeric vector of length two giving the lower and upper bounds
of the response support.}

\item{mu0}{Numeric target mean used to identify the reference distribution.}

\item{init}{List of initial values for the MCMC algorithm.}

\item{cdpglmControl}{Object of class \code{"cdpglmControl"} containing MCMC,
tuning, and copula-control parameters.}

\item{thetaControl}{Object of class \code{"thetaControl"} containing control
arguments for the theta update procedure.}
}
\value{
A list containing posterior samples, acceptance rates, prior settings,
support information, copula settings, and model specification information
used by \code{\link{cdpglm}}.
}
\description{
Performs MCMC to fit the CDPGLM model.
}
\keyword{internal}
