% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirglmFit.R
\name{dirglmFit}
\alias{dirglmFit}
\title{Main MCMC Fitting Function for the Dir-GLM}
\usage{
dirglmFit(
  formula,
  data,
  X,
  y,
  link,
  mu0,
  spt,
  init,
  dirglmControl,
  thetaControl
)
}
\arguments{
\item{formula}{Model formula supplied to \code{\link{dirglm}}.}

\item{data}{Data frame containing the variables appearing in \code{formula}.}

\item{X}{Numeric design matrix.}

\item{y}{Numeric response vector.}

\item{link}{Link object containing \code{linkfun}, \code{linkinv},
and \code{mu.eta}.}

\item{mu0}{Numeric target mean used to identify the reference distribution.}

\item{spt}{Numeric vector giving the support of the response variable.}

\item{init}{List of initial values for the MCMC algorithm.}

\item{dirglmControl}{Object of class \code{"dirglmControl"} containing
MCMC and tuning parameters.}

\item{thetaControl}{Object of class \code{"thetaControl"} containing
control arguments for the theta update.}
}
\value{
A list containing posterior samples, acceptance rates, prior settings,
and model specification information used by \code{\link{dirglm}}.
}
\description{
Internal function used by \code{\link{dirglm}}.
Performs MCMC to update model parameters.
}
\keyword{internal}
