% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpglm.R
\name{dpglm}
\alias{dpglm}
\title{Fits a DP-GLM model}
\usage{
dpglm(
  formula,
  data = NULL,
  link = NULL,
  dpglmControl = dpglm.control(),
  thetaControl = theta.control()
)
}
\arguments{
\item{formula}{An object of class  \code{\link[stats]{formula}}.}

\item{data}{An optional data frame containing the variables in the model.}

\item{link}{Optional character string or link object specifying the mean link
function. If \code{NULL}, a default link is chosen based on the response
support: \code{"logit"} for responses in \code{(0, 1)}, \code{"log"} for
nonnegative responses, and \code{"identity"} otherwise.}

\item{dpglmControl}{Optional control arguments.
Passed as an object of class "dpglmControl", which is constructed by the
\code{dpglm.control} function.
See \code{dpglm.control} documentation for details.}

\item{thetaControl}{Optional control arguments for the theta update procedure.
Passed as an object of class "thetaControl", which is constructed by the
\code{theta.control} function.
See \code{theta.control} documentation for details.}
}
\value{
List of class "dpglm" containing information including
posterior samples, Metropolis-Hastings acceptance rates, and model specification.
See details for more information.
}
\description{
Fits a DP-GLM model
}
\details{
The "dpglm" class is a list of the following items.
\itemize{
\item \code{samples} A list containing the MCMC samples for
\code{beta} and \code{crm}.
\item \code{mb} Prior mean for \code{beta}.
\item \code{Sb} Prior variance-covariance matrix for \code{beta}.
\item \code{formula} Model formula.
\item \code{data} Model data frame.
\item \code{link} Link function. If a character string was passed to the
\code{link} argument, then this will be an object of class "link-glm".
Otherwise, it will be the list of three functions passed to the \code{link} argument.
\item \code{spt} Support of response variable \code{y} used in model fitting.
\item \code{mu0} Mean of the reference distribution obtained by model fit.
\item \code{burnin} The number of burn-in MCMC iterations.
\item \code{thin} Thinning parameter used for MCMC
\item \code{save} The number of saved MCMC iterations.
\item \code{beta_acceptance} Proportion of accepted proposals for beta during MCMC.
\item \code{crm_acceptance} Proportion of accepted proposals for CRM during MCMC.
}
}
\examples{
\dontrun{
set.seed(123)

n <- 50
x <- rnorm(n)
y <- 1 + 0.5 * x + rnorm(n, sd = 0.25)

dat <- data.frame(y = y, x = x)

fit <- dpglm(
  y ~ x,
  data = dat,
  link = "identity",
  dpglmControl = dpglm.control(
    burnin = 10,
    thin = 1,
    save = 50,
    spt = range(y),
    seed = 123
  )
 )
}

}
