% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{f0_update}
\alias{f0_update}
\title{Function for updating f0}
\usage{
f0_update(
  y,
  spt,
  cr_f0,
  cr_f0y,
  cr_dir_parm,
  cr_mu,
  cr_tht,
  cr_bpr2,
  cr_btht,
  dir_pr_parm,
  ind_mt
)
}
\arguments{
\item{y}{response variable}

\item{spt}{y support}

\item{cr_f0}{current f0 = \{f0(s_j), j=1(1)l\}}

\item{cr_f0y}{current f0(y) = \{f0(y_i), i=1(1)n\}}

\item{cr_dir_parm}{current dirichlet proposal parameter}

\item{cr_mu}{current mu}

\item{cr_tht}{current theta}

\item{cr_bpr2}{current \{second derivative of b(theta_i), i=1(1)n\}}

\item{cr_btht}{current b(theta)}

\item{dir_pr_parm}{dirichlet prior parameter}

\item{ind_mt}{indicator matrix with (i,j)th element: 1(y_i == s_j)}
}
\description{
Function for updating f0
}
\keyword{internal}
