% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{f0y}
\alias{f0y}
\title{Function for calculating \{f0(y_i), i = 1(1)n\}}
\usage{
f0y(y, spt, f0)
}
\arguments{
\item{y}{(vector) response variable}

\item{spt}{(vector) support, \{s_j, j = 1(1)l\}, of response variable y}

\item{f0}{(vector) \{f0(s_j), j=1(1)l\}: free parameter, f0(.) centering density}
}
\value{
(vector) \{f0(y_i), i=1(1)n\}
}
\description{
Function for calculating \{f0(y_i), i = 1(1)n\}
}
\keyword{internal}
