% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{lDir}
\alias{lDir}
\title{Function for log Dir pdf for x ~ Dir(a)}
\usage{
lDir(x, a)
}
\arguments{
\item{x}{(vector) value at which to evaluate pdf}

\item{a}{(vector) Dirichlet parameter}
}
\value{
(vector) log Dirichlet pdf
}
\description{
Function for log Dir pdf for x ~ Dir(a)
}
\keyword{internal}
