% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dirglm.R
\name{plot_dirglm}
\alias{plot_dirglm}
\title{Plot DirGLM Posterior Distributions}
\usage{
plot_dirglm(
  fit,
  what = c("beta", "f0"),
  pars = NULL,
  top_k = 6,
  order = c("magnitude", "variance", "none"),
  robust = FALSE,
  prob = 0.95,
  dens_args = list(adjust = 1, n = 512, na.rm = TRUE),
  common_xlim = FALSE,
  ncol = NULL,
  col = par("col"),
  band_alpha = 0.25,
  line_lwd = 2,
  point_pch = 16,
  point_cex = 0.8,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{fit}{A fitted \code{dirglm} object.}

\item{what}{Specifies which posterior to plot. One of \code{"beta"}
or \code{"f0"}. Defaults to \code{"beta"}.}

\item{pars}{Specifies the names or integer indices of
the \code{beta} coefficients to plot. See details for more information.}

\item{top_k}{The number of \code{beta} coefficients to plot if \code{pars} is \code{NULL}
and the number of coefficients is large. Defaults to 6.}

\item{order}{Specifies the sorting mechanism for \code{beta} coefficients.
One of \code{"magnitude"}, \code{"variance"}, or \code{"none"}. Defaults to \code{"magnitude"}.}

\item{robust}{Logical indicating whether to plot robust statistics for
\code{Estimate} and \code{Est.Error}. Defaults to \code{FALSE}.}

\item{prob}{Value between 0 and 1 indicating the desired probability
level for the credible intervals. Defaults to 0.95.}

\item{dens_args}{A list of additional arguments passed to the
\code{stats::density} function for plotting the posterior density of \code{beta}.}

\item{common_xlim}{Logical indicating whether to share the x-axis limits
across all \code{beta} panels. Defaults to \code{FALSE}.}

\item{ncol}{Number of columns in the layout for \code{beta} panels.}

\item{col}{A string specifying the desired plot color.}

\item{band_alpha}{Specifies the transparency of the CrI band. Defaults to 0.25.}

\item{line_lwd}{Line width for the posterior density lines. Defaults to 2.}

\item{point_pch}{Point style for the \code{f0} plots.}

\item{point_cex}{Point size for the \code{f0} plots.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{main}{Plot title.}

\item{xlim}{Limits for the x-axis.}

\item{ylim}{Limits for the y-axis.}

\item{...}{Additional arguments passed to the \code{plot} function.}
}
\value{
An invisible list containing the selected coefficients or the
f0 support, depending on the value of \code{what}.
}
\description{
Plot DirGLM Posterior Distributions
}
\details{
If \code{pars} is \code{NULL} and the number of \code{beta} coefficients is greater
than \code{top_k}, the top \code{top_k} coefficients ordered by \code{order}
are plotted. The default for \code{top_k} is 6 and the default sorting mechanism
is the coefficient magnitude. If the number of coefficients is less than \code{top_k},
all of them are plotted.
}
