% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dpglm.R
\name{plot_dpglm}
\alias{plot_dpglm}
\title{Plot DPGLM Posterior Distributions}
\usage{
plot_dpglm(
  object,
  what = c("beta", "crm"),
  pars = NULL,
  top_k = 6,
  order = c("magnitude", "variance", "none"),
  robust = FALSE,
  prob = 0.95,
  dens_args = list(adjust = 1, n = 512, na.rm = TRUE),
  common_xlim = FALSE,
  ncol = NULL,
  grid_length = 200,
  col = graphics::par("col"),
  band_alpha = 0.25,
  line_lwd = 2,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  xlim = NULL,
  ylim = NULL,
  ...
)
}
\arguments{
\item{object}{A fitted \code{dpglm} object.}

\item{what}{Specifies which posterior to plot. One of \code{"beta"}
or \code{"crm"}. Defaults to \code{"beta"}.}

\item{pars}{Specifies the names or integer indices of
the \code{beta} coefficients to plot. See details for more information.}

\item{top_k}{The number of \code{beta} coefficients to plot if \code{pars}
is \code{NULL} and the number of coefficients is large. Defaults to 6.}

\item{order}{Specifies the sorting mechanism for \code{beta} coefficients.
One of \code{"magnitude"}, \code{"variance"}, or \code{"none"}.}

\item{robust}{Logical indicating whether to plot robust statistics for
\code{Estimate} and \code{Est.Error}. Defaults to \code{FALSE}.}

\item{prob}{Value between 0 and 1 indicating the desired probability
level for the credible intervals. Defaults to 0.95.}

\item{dens_args}{A list of additional arguments passed to
\code{stats::density}.}

\item{common_xlim}{Logical indicating whether to share x-axis limits across
beta panels. Defaults to \code{FALSE}.}

\item{ncol}{Number of columns in the layout for beta panels.}

\item{grid_length}{Number of grid points used when plotting the posterior CRM.}

\item{col}{Plot color.}

\item{band_alpha}{Transparency of the credible interval band.}

\item{line_lwd}{Line width.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{main}{Plot title.}

\item{xlim}{Limits for the x-axis.}

\item{ylim}{Limits for the y-axis.}

\item{...}{Additional arguments passed to plotting functions.}
}
\value{
An invisible list containing information about the plotted object.
}
\description{
Plot DPGLM Posterior Distributions
}
\details{
If \code{pars = NULL} and the number of \code{beta} coefficients is greater
than \code{top_k}, the top \code{top_k} coefficients ordered by \code{order}
are plotted. The default for \code{top_k} is 6 and the default sorting mechanism
is the coefficient magnitude. If the number of coefficients is less than \code{top_k},
all of them are plotted.
}
