% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.dirglm}
\alias{print.summary.dirglm}
\title{Print a summary for a \code{summary.dirglm} object}
\usage{
\method{print}{summary.dirglm}(x, digits = 3, ...)
}
\arguments{
\item{x}{\code{summary.dirglm} object}

\item{digits}{The number of digits to print. Defaults to 3.}

\item{...}{Additional arguments passed to \code{print}.}
}
\description{
Print a summary for a \code{summary.dirglm} object
}
\seealso{
\code{\link{summary.dirglm}}
}
