% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.dpglm}
\alias{print.summary.dpglm}
\title{Print a summary for a \code{summary.dpglm} object}
\usage{
\method{print}{summary.dpglm}(x, digits = 3, ...)
}
\arguments{
\item{x}{\code{summary.dpglm} object.}

\item{digits}{The number of digits to print. Defaults to 3.}

\item{...}{Additional arguments passed to \code{print}.}
}
\description{
Print a summary for a \code{summary.dpglm} object
}
