% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_helpers.R
\name{Sigma_beta}
\alias{Sigma_beta}
\title{Function for finding variance-covariance matrix of beta}
\usage{
Sigma_beta(X, mu, bpr2, rho, linkfun, mu.eta)
}
\arguments{
\item{X}{design (matrix)}

\item{mu}{(vector) \{E(y_i | x_i), i=1(1)n\}}

\item{bpr2}{(vector) \{second derivative of b(theta_i), i=1(1)n\}}

\item{rho}{MCMC update step size, a (scalar) in (0, 1]}
}
\value{
variance-covariance (matrix) of beta
}
\description{
Function for finding variance-covariance matrix of beta
}
\keyword{internal}
