% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.cdpglm}
\alias{summary.cdpglm}
\title{Creates a summary of a fitted \code{cdpglm} object}
\usage{
\method{summary}{cdpglm}(object, prob = 0.95, robust = FALSE, ...)
}
\arguments{
\item{object}{A fitted \code{cdpglm} object.}

\item{prob}{Value between 0 and 1 indicating the desired probability
level for the credible intervals. Defaults to 0.95.}

\item{robust}{Logical indicating whether to show robust statistics for
\code{Estimate} and \code{Est.Error}. Defaults to \code{FALSE}.}

\item{...}{Additional arguments, currently unused.}
}
\value{
An object of class \code{summary.cdpglm}.
}
\description{
Creates a summary of a fitted \code{cdpglm} object
}
