% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirglmFit.R
\name{dirglm.control}
\alias{dirglm.control}
\title{Control arguments for the \code{dirglm} algorithm.}
\usage{
dirglm.control(
  burnin = 100,
  thin = 10,
  save = 1000,
  rho = 0.1,
  gamma = 1,
  mu0 = NULL,
  spt = NULL,
  betaStart = NULL,
  f0Start = NULL,
  joint.update = TRUE,
  seed = NULL
)
}
\arguments{
\item{burnin}{Number of burn-in MCMC iterations. Defaults to 100.}

\item{thin}{Factor by which to thin MCMC iterations. Defaults to 10.}

\item{save}{Number of MCMC samples to return. Defaults to 1000.}

\item{rho}{MCMC update step size. A scalar in \eqn{(0, 1]}. Defaults to 0.1.}

\item{gamma}{Shrinkage parameter for the prior variance on \code{beta}. Defaults to 1.}

\item{mu0}{Mean of the reference distribution. The reference distribution is
not unique unless its mean is restricted to a specific value. This value can
be any number within the range of observed values, but values near the boundary
may cause numerical instability. This is an optional argument with \code{mean(y)}
being the default value.}

\item{spt}{Theoretical support of the response variable.}

\item{betaStart}{Initial value for the regression coefficients \code{beta}.
Defaults to the output obtained by fitting \code{gldrm}.}

\item{f0Start}{Initial value for the reference distribution \code{f0}.
Defaults to the output obtained by fitting \code{gldrm}.}

\item{joint.update}{Logical indicating whether to update \code{beta} jointly.
Defaults to \code{TRUE}.}

\item{seed}{Random seed. Defaults to NULL.}
}
\value{
Object of S3 class "dirglmControl"
}
\description{
This function returns control arguments for the \code{dirglm} algorithm.
Each argument has a default value, which will be used unless a different
value is provided by the user.
}
