% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirglmFit.R
\name{dirglmFit}
\alias{dirglmFit}
\title{Main MCMC function}
\usage{
dirglmFit(
  formula,
  data,
  X,
  y,
  link,
  mb,
  Sb,
  dir_pr_parm,
  mu0,
  spt,
  init,
  dirglmControl,
  thetaControl
)
}
\description{
This function is called by the main \code{dirglm} function.
}
\keyword{internal}
