% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dirglm}
\alias{summary.dirglm}
\title{Creates a summary of a fitted \code{dirglm} object}
\usage{
\method{summary}{dirglm}(fit, prob = 0.95)
}
\arguments{
\item{fit}{A fitted \code{dirglm} object}

\item{prob}{value between 0 and 1 indicating the desired probability
level for the credible intervals. Defaults to 0.95.}
}
\value{
A an object of class \code{summary.dirglm}. A list
containing with two elements, \code{beta} and \code{f0},
each containing a data frame with columns \code{Estimate},
\code{Est.Error}, \code{l-95\% Ci}, \code{u-95\% Ci}
}
\description{
Creates a summary of a fitted \code{dirglm} object
}
